+++ NEWS.md
@@ -1,3 +1,11 @@
+# semlbci 0.0.0.4
+
+* (Work in progress .... )
+
+* Can find the LBCI for the free and user-defined parameters in the standardized solution.
+
+* Starting values based on Wald CIs are uesd also for standardized solution.
+
 # semlbci 0.0.0.3
 
 * Can find the LBCI for free parameters and user-defined parameters.